+++ analyzer/testdata/src/findings/feature_enterprise.go
@@ -1,0 +1,4 @@
+// want +1 "custom build tag \"enterprise\""
+//go:build enterprise
+
+package findings