+++ analyzer/testdata/src/findings/findings_test.go
@@ -1,0 +1,5 @@
+package findings
+
+import "reflect"
+
+var _ reflect.Type